requirements.txt

fastapi
uvicorn
pydantic
openai
python-dotenv
jinja2
